requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.1
SQLAlchemy==1.4.41
Werkzeug==2.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
flask==2.2.2
flask_sqlalchemy==2.5.1
greenlet==1.1.3
idna==3.4
itsdangerous==2.1.2
requests==2.28.1
urllib3==1.26.12
ics==0.7.2
discord.py==2.6.3
